Handle collection submissions in view/submit.php (bug #786389)
Adds a collection parameter to view/submit.php, which causes an entire
collection to be submitted rather than an individual view.  To achieve
this, the db-specific parts of View::submit() are separated out into
another function View::_db_submit(), which takes a set of viewids, and
which is called by both the old View::submit() and the new
Collection::submit().

Change-Id: If66abd5bcab123366d2a747b763e2b6c9c638cc4
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>