Remove use of bitwise assignments on boolean values
This patch removes some inapproprate uses of the |= and &= operators
on boolean values.  On view/edit.php, &= was also used on the integer
return value from strlen as if it were a boolean, resulting in the
clean url field being uneditable in some instances where it should
have been editable (see bug #1006634).

Change-Id: I96a72c407309ed82a45d3130cffa06418be05fe7
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>