Use subdomain URLs on Settings and Edit view pages (bug #1006634)
When user subdomains are enabled, display the Profile url field on the
Settings page where it belongs: between the scheme part of the url and
the rest of the wwwroot.

On Edit Title & Description for profile pages, ensure the correct form
of the clean url is used by calling the appropriate function instead
of building the url.

Change-Id: Id4924165e7bf803fce46cf877a980d0c235176eb
Signed-off-by: Richard Mansfield <richard.mansfield@catalyst.net.nz>