Fix regression with mobile upload token (Bug #1057878)
New users were getting a token of "Array" set by default
when their settings were populated. It was useless.

Tokens were not being updated on the website. This was
due to changes in the api which required the old token
be passed when refreshing happened, and it was not
being passed in the json reply.

Change-Id: Ie8425e439b0b59134825c7922cfa887e7ad49c8b
Signed-off-by: Melissa Draper <melissa@catalyst.net.nz>