Fix the rss image exceptions preventing updating (Bug #1081431)
Check the image data is a string (eg url) or array before
proceeding, and set data to null if not.

Serialising is the only way to check if a value can be unserialised
so @ is suppressing the warnings thrown by non-unserialisable data.

Prevent the issue happening in future installs, and fix current
installs.

Change-Id: I05c88f95e2f511e491bcbd5fa40d3c41949eff7f
Signed-off-by: Melissa Draper <melissa@catalyst.net.nz>