Bug #1079498: Fix XSS in pagination URL
Currently, the url of a pagination (used for the prev/next links as
well as the numbered pages, and also the POST action in the form tag
used for selecting a variable limit, added in the commit listed below)
was not santized on output. This was discovered from the group member
search page which passed in the query as a GET paramter in the URL for
the pages. This uses slightly different code to some of the newer
paginations, but it may affect other places that use similar era
pagination setup.

The commit introducing the new selector for a variable limit was
 f3162f804855a3cdfa560339028da7a29560a491

This patch fixes this by sanitizing the url on output, in both the form
tag and the prev/next and numbered links.

CVE-2012-2253
Signed-off-by: Hugh Davenport <hugh@catalyst.net.nz>
Change-Id: Id9ed08ef5e61b12580e28f4b18975b2c409b594d