Bug #1079498: Fix XSS in pagination URL
Currently, the url of a pagination (used for the prev/next links as
well as the numbered pages) was not santized on output.  This was
discovered from the group member search page which passed in the query
as a GET paramter in the URL for the pages. This required that the
search query returns more results than the limit (10) to trigger the
bug. This also may affect other places that use a pagination setup.

This patch fixes this by sanitizing the url on output, in prev/next and
numbered links.

CVE-2012-2253
Signed-off-by: Hugh Davenport <hugh@catalyst.net.nz>
Change-Id: Id9ed08ef5e61b12580e28f4b18975b2c409b594d