License metadata - allow non-form-specific element help texts. (Bug #1095499)
Change the help text handling so that when a help text is not found for a
specific form/element combination, it also looks for a wildcard ANY/element
help text.

This is used by the license metadata feature to provide uniform help text for
the license fields on all forms where they appear.

Change-Id: I4c3f1a38d81ae54081e555b2296d28955329cf77
Signed-off-by: Jiri Baum <jiri@catalyst-au.net>