License metadata - pieform select element "allow other" option. (Bug #1095499)
Extend the Pieform select element with an "allow other" option. When this
option is present and true, the user can either select one of the options given
or enter a value in a text field.

This is used by the license metadata feature primarily to implement the "custom
license" option, where users can either select one of the configured licenses
or enter a URL. It is also used on the admin license edit page when there are
licenses in use which have not been configured.

Change-Id: I4e68651c4ea35699e9037ea8c9afca0bdbf94d11
Signed-off-by: Jiri Baum <jiri@catalyst-au.net>