Masquerading auditability - ask for reason, notify, emit event. (Bug #900983)
Masquerading (aka Loginas) is a useful and sometimes indispensable function.
However, previously it was rather too powerful, because admins can do anything
as the target user, with no indication that it is not the user themselves doing
so.

This feature adds some auditability to masquerading, by logging and reporting
who, when, why and (partially) what, as well as notifying the affected user.

See also bugs: #900983 and #1027574

This commit changes the masquerading feature itself to request a reason from
the admin masquerading, notify the user, and emit an event (for logging by the
next commit).

Change-Id: I066e9fdeb4d2e00679b2aa9b0b839cb4b78629a8
Signed-off-by: Jiri Baum <jiri@catalyst-au.net>