(Bug #1158086) Error when deleting the child of a deleted post
After you delete, we normally redirect you to the parent of the
post you just deleted. This code change makes it check whether
the parent is deleted and fall back to the nearest un-deleted
ancestor, or just leave the post id off entirely if it can't find
one (in order to fail more gracefully).

Change-Id: I7422b3d5b466ee55899b4959dd489dd76a8d5939
Signed-off-by: Aaron Wells <aaronw@catalyst.net.nz>