Leap2a: Add support for standalone Notes and their Attachments Bug#1117237
The exportation and importation of attachments for text boxes should
be working correctly.

The exportation of attachment files for resume options should be working
- however the leap2a dat afor them is not quite correct

The leap2a importation should upload the resume attachment files and
place them in Contens -> files.
- however it is not currently attaching them to the resume entry, the
  artefact_attachment table is not getting updated correctly.

TO DO: get the <link rel="related" href="portfolio:artefact?"/> to
be in correct part of xml file and also be able to save the item id to
the table as well

Change-Id: Ibda0fcc85265c458a59738f2c69e712b7324d445
Signed-off-by: Aaron Wells <aaronw@catalyst.net.nz>