Prevent non-helpful sql_mode error when trying to connect to a UTF16 DB
Bug 1184450. I don't know why this works but it does. When attempting to
connect to a UTF16 MySQL DB, if you first set the connection's character set
to UTF8 and then try to set the SQL_MODE to POSTGRES, it errors out on trying
to set the SQL_MODE. But if you try to set the SQL_MODE first, and *then* set
the character set to UTF8, it works fine.

Change-Id: I3528a92662708c081b15a111c625f0559fbd1c04