Edit forum post error if parent is deleted (Bug 1204309)
For Edit post, we can use the parent post to get the group or the topic details
even if the parent has the deleted flag set to 1. I am checking that the parent
is not deleted only for add post.

Since the get parent query is common to both add and edit post, I changed it to
search for a parent ignoring the deleted flag.

Change-Id: Ie0602d940d3c3b97e97159f8bfa540128c7f1f1d
Signed-off-by: Nitish Bezzala <nbezzala@yahoo.com>