Display of files in folders does not follow alphanumberic conventions (Bug 1180775)
The sort function adds a 0 in front of folders and a 1 in front of file
names to display folders first. So, the sort order is really by numbers,
and if a file has a number in the beginning, it ends up at the end of
the list.

I am separating the initial number and the file name with an 'a'
character, so that the sort order is - first folders, then files, and
all sorted by title.

Change-Id: I1367ebeaf030582beaac1892799dc844ecb51407
Signed-off-by: Nitish Bezzala <nbezzala@yahoo.com>