New watchlistnotification Plugin (Bug 1041228)
Introducing a new plugin watchlistnotification that responds to the
events saveartefact, blockinstancecommit and deleteblockinstance. It
stores the changed view and the blockinstance in a table watchlist_queue
and checks via cron if there were any changes on a view and if for that
view the last change has happened some time ago (the minutes are stored
in config under watchlistnotification_delay, the default is 20min).

If so, a message is generated that informs the watchlist recipient about
which view and which block-instances on this view have been touched
(added or changed).

As there is no way to disable the built-in/old watchlist-notification-
system, this is disabled in the mahara-core code, that is,
artefact/lib.php and lib/view.php

Change-Id: I039c5285cdd1b09ed9eb38a647e0c1510c3cabb9
Signed-off-by: Tobias Zeuch <tobias.zeuch@kit.edu>