Creation of new blocktype 'comment' (bug #1245679)
- allow the feedback to display in a block.

Things to note:
1) The feedback block is only to display view's comments on the
view/view.php page and NOT on view/artefact.php page.

2) When in edit mode (view/blocks.php) the feedback block just
displays a 'holder' message

3) When feedback block is on page the feedback does not display at
base of page anymore and when one clicks the 'Place feedback' link the
feedback form pops up over everything similar to how editing a block
works.

4) As the feedback block only works/displays on the view page I have
not fully integrated all the stuff that feedback creates ie js/form
into the blocktype. I have left the view/view.php to do the js and
render form - and simply position the form on the page with js.

This means that if js is off the edit form will still be in it's
old place. I felt this was ok because if js is off and the form was
in the block it could mean things been too squashy if there was lots
of columns.

Change-Id: I1edaa886c5a0df1f6feb24bcdd6d8878a833d673