Problems with image gallery and file/folder selection Bug 123530 and Bug 1236122
When js slideshow was chosen on a page without a slideshow block
(eg a new page) the js code returned by the selection was not loaded
fast enough

- I've altered it so the js is loaded on opening of config block

- Also folder was not staying selected when re-editing the block
due to it being not ArtefactTypeFile - have adjusted the check to
include ArtefactTypeFolder as well

- The switching between My files and Site files was breaking due to
folder artefact chooser not having certain elements on the page that the
js was trying to access. Have made it so js checks for elements before
trying to update them.

Change-Id: I9b8cbe786b0d35109f05a664f04665e9a9967d92
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>