Allowing a theme to disable skins from being viewed (bug 1237013)
To get a theme to stop showing skins on portfolio pages add this
to the theme's themeconfig.php
$theme->skins = false;

When you go to add a skin to a page it will warn you that skins are
incompatible with this theme - you still can add a skin to the page
but it will not show up when you view the page.

Also if 'Page themes' are turned on it will warn you that the chosen
page theme for the page is not compatible with skins.

Change-Id: I694bf14dffddd12373c4079b57c047fa509ab7b7
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>