Improve leap2a interactive import's handling of resume composite artefacts
Bug1237791

Also adding a new bit of API: If an $entrydata being passed in to PluginImport
includes a 'defaultdecision' field, we use that instead of the normal logic
to decide which non-ignore decision to use.

ArtefactTypeResumeComposite artefacts are now allowed only the "Append" option,
which adds new composite rows via the "ensure_composite_value()" function.

ArtefactTypePersonalInformation artefacts continue to use much the same logic
as we were previously (erroneously) using for ArtefactTypeResumeComposite artefacts.
Note that in non-interactive import, they'd be populated using the import_author_data()
function, but that doesn't work in interactive import because it would overwrite everything.

Change-Id: I39304d53150cc6e8a880b2c2b68cfff1afb45349
Signed-off-by: Aaron Wells <aaronw@catalyst.net.nz>