Allowing file browser to stay open when changing dirs (bug 1240415)
Problems here is there are 2 file pickers on the same form so we can't
use the normal js success callback to update form so we have to go the
non js way - but if we have browse=1 in the url the form won't save.

Patch looks to see if folder is set in url and if so then we
are in a browse mode.

Also adjusted the template to allow for the adding of the 'select
a file' button (in hidden state) to the page when entering the
subdirectories. this will mean the hidden button will be added to a few
places where it is not need (eg selecting an image gallery
blockinstance) but as it is hidden by default no one will see it.

A better fix would be to get multiple filechooser elements working in one form
but this isn't likely to happen at the moment.

Change-Id: Iddbd677a81d216f71cdeae86a02ed107261c04b5
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>