Max-width being too big for the column (bug 799441)
When you set the width of the slideshow in the config (which defaults
to 400) and try and place it in a thin column and the images in the
folder you select are quite large then they can end up hiding the
right arrow.

I've added asome js to allow for this. After page fully loads it adjusts
the max-width setting on the images to fit with the allowed space
of the column when the column width + slideshow arrows are less than the
specified max-with from the blockinstance's width setting

Change-Id: I111b1b51d2583a9c80368b7cbd39bc3c2f6c7df4
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>