Revert the upgrade section that searches for orphaned group content
Bug1236661: It turned out this process can take a very long time
for substantially sized Maharas, which are the ones that are most
likely to have been around since 1.4 and benefit from this. Since
we've updated the code where the orphaned pages were causing problems,
this bit is no longer strictly necessary, so it's best just to
remove it.

Change-Id: I628eb24fd7eb38a50b484df16053c55d7264d89b