New plugin as user-message system to multiple recipients
Bug 1246702: This is a plugin that replaces the current user notifications with a
message system for multiple recipients.

- The plugin redirects calls to sendmessage.php and inbox.php.
- New Usermessages are stored in a separate table from the userrelation, where there
  is also kept track of whether the notification has been read or deleted.
- Messages aren't deleted from the database, until all participantas (sender and all
  recievers) delete it.
- The plugin comes with an outbox that shows messages that have been sent by the
  logged in user.
- Both inbox and outbox show new messages and old messages alike.

Note that the code is at pre-release state:
The message-count per user is kept up to date with database-triggers, as is done for
the old system.
For this new system, you might not want to install it on a productive system or it
might mess up your message count when you deactivate the plugin.

Change-Id: I586af86266b5fb0e2d676bd1401b7d879d0122ef
Signed-off-by: Tobias Zeuch <tobias.zeuch@kic-innoenergy.com>