Artefacts still accessible when disabled/hidden (bug #646691)
Currently we only hide links to artefact when we 'hide' them on the
extensions page rather than have them not accessible.

This patch will check to see if artefact is active and if it isn't
will display a access denied message.

note for testing: when hiding an artefact plugin it turns off the blocktypes
associated with it - but when you re-enable the artefact it does not
re-enable the blocktypes

Change-Id: Ic2282ededa7e85696dcaaac0768d35c52ecf3d01
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>