Embed pdf has horiz scrollbar on default 3 column layout (bug 1237137)
Seen as the 3 column layout is the default one the embed pdf view
should fit within it.

Currently if you have a browser windo open fully, a 3 column layout,
and a pdf embedded you get a horizontal scrollbar on the embed pdf.

A slight adjustment of the min-width and it goes away.

The scrollbar will be there for 4 or 5 column widths but that is ok as
they are quite thin columns.

Change-Id: If03f4ce148e90806eb06c901b43671f30cdf8493
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>