Embed pdf has horiz scrollbar on default 3 column layout (bug 1237137)
Seen as the 3 column layout is the default one the embed pdf view
should fit within it.

Currently if you have a browser windo open fully, a 3 column layout,
and a pdf embedded you get a horizontal scrollbar on the embed pdf.

A slight adjustment of the min-width and it goes away.

The scrollbar will be there for 4 or 5 column widths but that is ok as
they are quite thin columns.

Change-Id: Ib01118d9fffab14440308eaf04b5038fdfeca9c3