Getting mysql not to use CAST() (bug #1081947)
In the 1.8+ codebase there are only two places where both mysql and
postgres use CAST(). So I've tried altering the calls so that mysql
doens't need to use CAST().

This will probably need to be tested with a large data mysql db site
to see if any speed gains are made, and also the calls are still
working correctly.

All other places it is being used by postgres only
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>

Change-Id: I9eaf196d062ef2a62bfdae0df179281f3411c392