Fill in the missing implementation for find_key_name()
Bug #1255378

This fundamentally changes the behavior of find_key_name(). Before, it only generated what the default name for the key would be, now it actually checks whether a key of that description exists in the DB and returns its actual name in the DB.

In core the method was only used by XMLDBGenerator in the methods for dropping or altering a key, so this change will actually make it more robust.

In theory, though, if there is third-party code relying on the current behavior of find_key_name(), they could have problems.

Change-Id: I0470422a6e6514af98beb83e884178ee7f341ccd