Allowing non-institution site admins to publish (bug #1255361)
Currently, if a site admin tries to edit an institution page and
attach a file to a block (eg image block) and they don't belong to
that institution, the process fails.

Have altered the can_publish_artefact() to allow those that can edit
institutions to be allowed to also publish.

Change-Id: I05a2160e84814add067019e4bde654a6ebd561b1
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>