Add forum objectionable content reporing functionality (bug #1024872).
The feature allows to report objectionable content in the forum posts and
topics. When post is reported, the notification is sent to site admins,
group admins and forum moderators. Any of above can take action and either
mark the post as not objectionable or delete it. In both cases the
notification about action will be sent to users who were originally
notified about objectionable content, so that they will be aware on other
person action and outcomes. Site admin normally can't access the content of
the forum he/she is not a member of, however in the case of objectionable
content, site admin will be able to have temporary rights similar to group
admins, making possible to delete or edit any post in the given forum. Once
the issues is resolved (forum no longer contains objectionable content),
admin will not longer be able to access forum in the group he/she is not a
member.

Change-Id: I12e459e2f754fcc7f5eeb0bad2f646927ad03cf8
Signed-off-by: Ruslan Kabalin <r.kabalin@lancaster.ac.uk>
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>