Allowing for users to not show up in public search (bug 1257953)
Thinking of privacy issues here - when should the users showup on
search results.

This patch allows users to show up based on the access of their
profile page (accesstypes: public, loggedin, friends) and whether
'Show users in public search' is set or not.

Normally all profile pages are accessible by logged in users - but
this can be altered if 'Logged-in profile access' is unchecked and
users remove the access off their profile page.

Change-Id: I4daa8cb2812bddb231ba489dfeefb4843b653d40
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>