Allow site_content to be institution specific (bug #1254299)
Changes include:
- added an institution column to the site_content table
- added an 'Edit site pages' page under Admin -> Institutions
that is accessibe by institution admins
- added an 'institution' option to the edit site pages form - this is
a hidden field if user can edit only one institution.

On upgrade it updates the site_content table to give current data the
institution on 'mahara' (incl. local site pages) and for each
institution it replicates the data already in the db for the default site (excl.
local site pages) so that every site has their own versions, which can
be adjusted as one sees fit.

On creation of new institution it creates the rows in site_content
table but with the default strings (like what you see when you first
install a mahara) but sets the sitepages column in institution table
to default (mahara). On deletion of institution it removes the rows in
site_content.

A user on login sees the institution site page based on what
institution theme they see.

On logout the 'lastinstitution' cookie is set allowing for them to see
institution specific site pages.

The 'No institution' (mahara) site pages can only be edited through
Configure site -> Edit site pages.

Also allow for an institution site page to be viewed if 'institution'
variable is passed to it eg terms.php?institution=testing allowing for
another way to access info when logged out.

Change-Id: I2ed30b63c15bf676d83eb2231f48c4ca23ce8b53
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>