In Resume, show composite description in cell rather than row (Bug #1145156)
Changed showhideComposite() to add the requested content (description and
attachments) to the cell containing the title of the composite, rather than
adding a new row to the table.
This means that in places such as the Education history table, clicking on the
qualification will open the description immediately below the title instead
of creating a separate row.

Change-Id: Id05bf73e2932108abd9979359bb7f64e6b3ca3fc
Signed-off-by: Jono Mingard <jonom@catalyst.net.nz>