Use focus management in Feedback form (Bug #1259764)
"Place feedback" button should now focus the Message input of the Feedback
form. Also, if the Feedback block is used on a page, the feedback form is
moved so that it's under the "Place feedback" button in the HTML - this
means that the document stays in a logical order for screen reader users.

Change-Id: Ieeb71bc7b25e48b7b69683173c01cf1659e4144d
Signed-off-by: Jono Mingard <jonom@catalyst.net.nz>