Replace obsolete TinyMCE GoogleSpell with in-browser spellcheck as default
Bug 1212541: The web service the TinyMCE GoogleSpell code was using,
is no longer in existence. This patch changes things so that TinyMCE will
let the browser's spellchecker work, by default.

The "pathtoaspell" is now null by default. If pathtoaspell is null, we
eliminate the TinyMCE spellcheck button and set the "gecko_spellcheck"
flag to let the browser spellcheck work. If pathtoaspell is not null,
we show the TinyMCE spellcheck button as before.

Change-Id: I9fd2b12169669d3d9705a1e23dc6c03af6f01948