Getting rid of the placeholder "Social" artefact plugin
Bug 1259538: We created a new artefact plugin called "social" to hold
three of the profile completion tasks. But I think this is too much
overhead just for that, so I'm folding them into the "internal" and
"comment" plugins.

(In the long run, we should probably make it possible for *any* plugin
type to add progress bar tasks, and then add these as local plugins or
something.)

Change-Id: I7d129069325f45409be0c73dedd6ddcc3db17489