Anonymous comments not showing the correct profile icon (Bug #1300289)
The anonymous comments should show the no_userphoto icon.

I've taken the patch supplied by Brian King and simplified it a bit so
that if no user is supplied to profile_icon_url it gives us the
no_userphoto icon rather than needing a new Dwoo function to handle
this.

Change-Id: I469fad4532b3ae91e1d673ec33d04d04c4fc3c7d
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>