Set default layout columns/widths hardcoded once (Bug #1307760)
And making sure that it exists in the db before using it.
This allows the fetching of the id of the default column widths to be
used in view layout page.

Also tidied up the js for the view layout page so it uses the value
from php rather than hardcoded string.

Also removed duplicate 'if (empty(self::$layoutcolumns)) { }'
statement from lib.view.php

Change-Id: Icf1b4292a256c5ff6d748c9e6046d6c71ea298b2
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>