Prevent global form error from incorrectly showing up on forms with button-submit functions
Bug 1311940: There's an obscure pieforms feature where a form can have multiple submit
buttons, and each submit button can have its own submit callback function named
{formname}_submit_{buttonname}. These functions are supposed to redirect at the end.
If they don't, due to a bug in the implementation of the global form message we currently show
an error message at the top of the form saying there was something wrong with the user's
submission. Instead, we should just let the form go through, and log a debug message for
the devs.

Change-Id: Ib0e9b7e2f58737c9c736bdd1c033bc020985a60b