Altering the toobar button layout to be more like moodle (Bug #1282341)
Now the 2nd / 3rd rows of buttons are hidden by default and are able
to be toggled on/off with the top left button.

Made the button more accessible by having it highlighted as on when
clicking to show all buttons. And also added different tooltip
messages depending on whether the button is in the on or off state.

Change-Id: Ia4b4f5631218088a4293abc991620c646b8d2d6c
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>
Signed-off-by: Yuliya Bozhko <yuliya.bozhko@totaralms.com>