Allowing the 'install font' to take a zip file (Bug #1312672)
When installing a normal font you have to select 4 or 5 different
files to be allowed to install the font - and most often you get these
files from 'fontSquirrel generator' in a zip file.

I've altered it so that you can either upload each file individually
or you can upload a zip file.

Change-Id: I381b2a06cf4e82c6c7bdb7b5413c620aa98eb467
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>