Allow collection's pages list to be added/edited by drag/drop (Bug #1320716)
You can now alter the sorting of the list by dragging the page name to
a new position.

You can also add in new pages to the collection by dragging them in
from the 'add pages to collection' list. If there are no existing
pages in collection you need to drop a page onto the 'No pages' grey
box.

Also added touch-punch js lib to allow for the drag/drop to work as
well on mobile devices.

Change-Id: Id1af83abc50b76a7a8d56da8767ed4ce0ef76f77
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>