Drop support for auth plugins using site-config functions for instance config
Bug 1331863: Now that we've add a default validate_config_options()
implementation to the base Plugin class, we get a warning when saving
the config for auth plugins that say yes to has_instance_config() but have
not implemented the (optional) function validate_instance_config_options.

This is because of backwards-compatibility code to deal with older
auth plugins that were using get_config_options(), validate_config_options(),
and save_config_options(), to handle instance config. We made this change
in Mahara 1.5 and added a warning message then, that says to update the
plugin. I think that's a long enough warning period to just go ahead and
drop support for it.

From now on, plugins that have instance configs, will have to implement
get_instance_config_options(), validate_instance_config_options(), and
save_instance_config_options(). (All the core auth plugins already do.)

Change-Id: Ia135a96a6d8c36d36708a9b9a66eaef71bc788e9