Searching by user in Shared Pages (bug 897586)
Include searching by a user to the Shared Pages tab.

This is a pretty complicated area as the View::view_search() is used in many areas.
Have not added a new search type of 'users only' like the 'tags only'.
Have only updated the general search type of 'Title, description, tags' to include users.

Change-Id: I6020aa5f432133649e50ea180bbd8536c8e30e72
Signed-off-by: Ghada El-Zoghbi <ghada@catalyst-au.net>