(Bug548021) Add support for anonymising pages.
This patch adds support for anonymising pages.

It creates a site option in the General Settings section named 'Allow
anonymous pages' that must be checked for any page to be anonymised.

If the site option is enabled, a new 'Anonymise' setting in the
'Edit title and description' tab of pages is enabled, allowing the owner
to request that this page be anonymised.

When both settings are enabled and a user views a page, two things may
happen to the author's information.

First, if the user is not logged in or does not have admin, staff or
owner privileges for the page, the author's name is replaced by the
anonymous text (defaulting to "(Author's name hidden)") in both the
body of the page and the metadata author field.

If the user does have admin, staff or owner privileges for the page,
the author's name is anonymised as above except that the text displayed
is made into a link. When this link is clicked, the anonymous author
text is replaced by the normal author information for the page.

Anonymous pages displayed in the 'Latest pages' block and shared with
a group are also anonymised by this patch.

Change-Id: I2e2c92f641329a1a305cf58a5c5d47bf95436a8b
Signed-off-by: Nigel Cunningham <nigelc@catalyst-au.net>