Copy secret URLs to clipboard button (Bug 1314890)
This patch adds a new button that allows the user to copy
secret URLs to the clipboard without having to manually
select the URL text (a process prone to error). Since
Javascript can't do this directly, it uses the ZeroClipboard
library's Flash based implementation, which adds an invisible
'movie' over the button.

The patch includes error handling, so that if Flash is disabled
or Zeroclipboard is unable to run for some other reason, the
buttons are automatically hidden. Note that if Adobe is set to
ask whether to allow flash to run, the buttons are enabled and
there's no way for us to detect if the user then denies permission.
Since they're explicitly refusing permission and there's nothing
we can do, I'm just documenting the fact here.

Change-Id: Idf6a408b77adc3f5401192688046d8aaa9e125ae
Signed-off-by: Nigel Cunningham <nigelc@catalyst-au.net>