User in multi institutions can lose progress bar (Bug #1347362)
If a user is in two (or more) institutions and the first institution has
profile progess set up for it and the other doesn't, the switching of
the select box makes the progress sideblock disappear.

This patch checks that if any institution the user belongs to has
profile progress then display the select box if they belong to two or
more institutions and show either the bar or a message mentioning this
intitution doesn't have progress.

To test:
- user with 1 institution that has progress
- user with 1 institution that has no progress
- user with 2 institution that both have progress
- user with 2 institution where one has progress
- user with 2 institution where none have progress

Change-Id: I4daca26b4b9cfef9faa73be34c34ad931d89465e
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>