Checking to make sure artefact/block library still exist (Bug #1319634)
Sometimes views will have artefacts/block instances on them but the
plugin has been deleted. This should not stop a user being deleted.

To test:
Add a plugin to the site (eg openbadgesdisplayer block plugin, or cpds
artefact plugin) and install it.
Add a page and add a block to the page.
Delete the dir from the site for the plugin
Try to bulk delete the user

Change-Id: Ib23ad79d9f1ffa0bca56ad7ccb6083333527e64c
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>