Getting IE to behave as Chrome/FF when clicking file icon (Bug #1348595)
Currently in IE/FF/Chrome if you tab through a list of files on
artefact/file.index.php and hit space when on the icon you are given a list of places
you can move the file to. (Note: have some files and some folders when
testing).

But when you click on the icon you only get the list when in FF/Chrome.

This patch should allow the clicking to also work in IE 10

Change-Id: I958d52220912169f1fc0198a47428e077aa6ed3e
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>