(Bug1364691) Add artefact_instances_from_ids function.
Add support for getting multiple artefacts from an array
of IDs, all at once. The main saving is that we then do
one SQL query for the whole array instead of one for each
element in the array.

In testing admin/users/bulkexport.php, this reduced
execution time for an export of 1000 users by 12 seconds
(approx 10 percent).

Change-Id: I51867cd4693a7572dd609b98b6503251e473078a
Signed-off-by: Nigel Cunningham <nigelc@catalyst-au.net>