Removed maildisabled and authinstnace options from list (Bug #1352027)
The way that bulk user csv upload works the authinstance is set for
all users on the csv via institution field in the upload form so we
don't need that column in the csv file.

Same with mail disabled - there is a checkbox on the form that we can
tick/untick to deal with that.

So I've got the system to not display those options on the choices list.

Change-Id: Ib2b3e2f19ee0f8aeed03c7dc04af076721ce8b41
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>