More Google maps urls not working in Google Apps block (Bug #1353759)
Was not working for urls with @ latitutude,longitude, zoom eg:
https://www.google.com/maps/@37.509345,-121.9596665,12z

And not working for shorthand ulrs like https://goo.gl/maps/pZTiA

So I've added a check for the urls with @ latitutude,longitude, zoom
and turned them into an acceptible embed url. And for shorthand urls
I've added a generic function to fetch the url the short one resolves
to and so it can parse that one instead of the short url.

Change-Id: I3c76d2608e9708f1722e243cdc02a43f1885b815
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>