Restore "toggle" button to TinyMCEs with "tinymceconfig" param
Bug 1381738. All the TinyMCE instances that use the "maharaimage" plugin are providing
their own tinymceconfig param which redefines the whole list of plugins in use. So they
didn't get the message when "tooltoggle" became one of the standard plugins.

This fix is a bit of a hack. A more robust fix would provide a way for a TinyMCE to
*add* to the list of plugins, or just specifically say it wants to use maharaimage
instead of image.

Change-Id: Ia8128e5d70bd4be1e5d4b45efae368a88174c4c5