Add a <meta name="generator"> tag with the Mahara major version number
Bug 1384497. This also adds a new $config->series to lib/version.php,
which stores the major version (or as we sometimes call it, "the series")
of the current release. It seemed better to store this separately rather
than to try to parse it from $config->release, because we may change
our release numbering strategy in the future.

Change-Id: If4c0d7352e9433cdaf54a03bb922356d6c950e7d