Bug1051844: Fix block configure dialog width.
The configure dialog for blocks is currently set to a fixed width via
Javascript, but there is no handler for browser resizing and it results in the
dialogue being very wide for the sake of just one line when a license agreement
is enabled.

Fix this issue by making all dimensions percentages, and using a combination of
max-width: 90% and width: max-content. In addition, make the dialog fixed
position with its own scrollbars as needed (the body's scrollbar is disabled
when the dialog is shown).

Contextual help's width code and CSS is also adjusted so that it wraps within
dialog boxes rather than pushing the width out.

Change-Id: If2a40739f1510cedcf9a6133ccfb8f0a18257356