Annotation artefact: Bug 1397759
A new artefact similar to the comment artefact but with less
functionality (i.e. no attached files, etc).
It's an explenation of why a particular evidence meets a
particular standard.

If an annotation is created and added to a page, when the user
deletes it from the page, the instance is deleted along with the
annotation and its feedback.

TODO:

1. Imports seem to be working.
Can get all comments to import and display.
Needs some serious testing.

2. Made changes for broken images but another bug was reported and is
currently being worked on. So, may not need the fixes in here. Changes in:
- htdocs/artefact/file/download.php

To completely fix the broken images for all artefacts, changes are also required
in htdocs/lib/embeddedimage.php to delete based on resourceid instead of fileid.

Change-Id: Ibdb2e1c6500862645bac741bf61cff37e5a5b35c