Display pre-1.10 TinyMCE smiley images
Bug 1408542. TinyMCE 4 renamed the smiley plugin from "emotions" to "emoticons",
and consequently the smiley graphics moved from .../emotions/img
to .../emoticons/img, and all the smileys from before the Mahara 1.10 upgrade
show up as broken links.

As a workaround to make this work without a DB upgrade, we put a second copy
of the smiley images in the old location. For the next major release, we should
do a DB upgrade script that locates all references to the old location and
rewrites them to be references to the new location.

Change-Id: If5b193f526a50b5e200681cc18d4188cd55104ab