Fix error with makefile with coding standards check
During a make push with one of the academy student patches, they got
the error http://paste.dollyfish.net.nz/ea833c with patch
https://reviews.mahara.org/#/c/4176/2

Going through, it looks like it was in the check for the coding
standards and they had a branch? called that for some reason. I'll
check in a bit when I verify.

Added -- to seperate git arguments from filenames to fix this.

Change-Id: I36b400cff115977ebb888ced57ba7fbbb3a3fcc0